add detail/parsing/binary_writer.hpp
