+++ Makefile
@@ -19,7 +19,8 @@ SRCS = ${SRCDIR}/json.hpp \
 			 ${SRCDIR}/detail/iterators/iteration_proxy.hpp \
 			 ${SRCDIR}/detail/iterators/json_reverse_iterator.hpp \
 			 ${SRCDIR}/detail/parsing/output_adapters.hpp \
-			 ${SRCDIR}/detail/parsing/binary_reader.hpp
+			 ${SRCDIR}/detail/parsing/binary_reader.hpp \
+			 ${SRCDIR}/detail/parsing/binary_writer.hpp
 
 # main target
 all:
